feat: Add TextView to display list from ViewModel
Adds a `TextView` with the ID `textViewList` to the `activity_main.xml` layout.

This new view is constrained below the `search_edit_text` and to the start of the parent layout. It is configured to display data from the ViewModel by setting its `android:text` attribute to `@{safeUnbox(viewModel.textList).toString()}`.